imp(calc winner): - moved out function to helper. - wrote tests - cleaned some logic.
