+++ src/components/board.jsx
@@ -7,35 +7,7 @@ import LobbyList from './lobby-list';
 import Square from './square';
 import GameStates from '../constants/game-states';
 
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  let filledSquares = 0;
-  for (let i = 0; i < squares.length; i++) {
-    if(squares[i]) {
-      filledSquares++;
-    }
-  }
-  if (filledSquares === squares.length) {
-    return 'draw';
-  } else {
-    return null;
-  }
-}
+import { calculateWinner } from '../helpers';
 
 export default class Board extends React.Component {
   constructor(props) {
